organize read datum
